Extract PasswordTextField to components
